requirements.txt
playwright
sqlalchemy
typer
python-dotenv
openai